move migration logic to its own module
