+++ homeassistant/components/zwave_js/helpers.py
@@ -3,7 +3,6 @@ from typing import List, Tuple, cast
 
 from zwave_js_server.client import Client as ZwaveClient
 from zwave_js_server.model.node import Node as ZwaveNode
-from zwave_js_server.model.value import Value as ZwaveValue
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant, callback
@@ -11,32 +10,6 @@ from homeassistant.helpers.device_registry import async_get as async_get_dev_reg
 from homeassistant.helpers.entity_registry import async_get as async_get_ent_reg
 
 from .const import DATA_CLIENT, DOMAIN
-
-
-@callback
-def get_old_value_ids(value: ZwaveValue) -> List[str]:
-    """Get old value IDs so we can migrate entity unique ID."""
-    value_ids = []
-    command_class = value.command_class
-    endpoint = value.endpoint or "00"
-    property_ = value.property_
-    property_key_name = value.property_key_name or "00"
-
-    value_ids.append(
-        f"{value.node.node_id}.{value.node.node_id}-{command_class}-{endpoint}-"
-        f"{property_}-{property_key_name}"
-    )
-
-    endpoint = "00" if value.endpoint is None else value.endpoint
-    property_key = "00" if value.property_key is None else value.property_key
-    property_key_name = value.property_key_name or "00"
-    value_id = (
-        f"{value.node.node_id}-{command_class}-{endpoint}-"
-        f"{property_}-{property_key}-{property_key_name}"
-    )
-    value_ids.extend([f"{value.node.node_id}.{value_id}", value_id])
-
-    return value_ids
 
 
 @callback
